Remove unused RBNF data.
I've left support for RBNF data in "icu_data_generator.py" for now.
I made a couple of changes to the script: I made it slightly more
clear what -h and -v do, and I changed the -v output so that it
doesn't look like we accidentally run the same genrb command four
times in a row (we actually do, but it's in a different directory
each time). This looked like a bug to a naive user (me) running the
script for the first time.

Everything still builds and boots. Not sure what other tests to do
to prove there isn't a rhinoceros in the room.