Update patch set 1
Patch Set 1:

I confirm that the code merges and builds fine in the open-source tree. There are conflicts with Google's internal master tree. The makefile conflict is easy to resolve, but the code conflict will probably require more knowledge of the underlying code to make sure that such exports are added to all relevant functions (as there might be some new ones).

Elliott, one you're confident that you'll be able to deal with that conflict (and when our internal merges are flowing smoothly), please mark this change as verified and submit it.

Patch-set: 1