MediaScanner crashed on Japanese meta file
MediaScanner crashed while scanning a file, which has a Japanese
filename and has no metadata or too long metadata. In this case,
the filename will be stored to DB as the title, and some Japanese
filename causes a crash while creating search keys.

When the data converted into the sorting key exceeds the size of
the buffer, buffer over run is generated by a bug in
ICU4C(ucol#ucol_calcSortKey).

In ucol#ucol_calcSortKey, the sorting key is set to the buffer
while advancing the pointer by one byte. When sorting key is
two-byte primary and compress data, the pointer is advanced too
much and a buffer over run is generated.

Change-Id: I595325c89633c44b08673411321be315d574409e