libicuuc: Fix memory corruption in libicuuc
The ISCI routines are making use of some predefined sets which gets
allocated and initialized in the ISCI open routine and freed in the
ISCI close routine.

These predefined sets are global, so if we use ISCI operations with
multiple nested contexts we can easily run into memory corruption
issues.

An example for a double free issue:

_ISCIIOpen (creates context 1) -> PNJ_CONSONANT_SET = malloc(...)
_ISCIIOpen (creates context 2) -> PNJ_CONSONANT_SET = malloc(...)
...
_ISCIIClose (context 1) -> free(PNJ_CONSONANT_SET)
_ISCIIClose (context 2) -> free(PNJ_CONSONANT_SET) --> double free

If multiple threads are used then we can also run into use before
initialization errors beside double free errors.

To fix this issue we allocate and initialize the global sets only if
this is the first open and we free the global sets only if this is the
last close.

Change-Id: Icf5eff3b302a30e590332437d594a0b60941192f
Author: Octavian Purdila <octavian.purdila@intel.com>
Signed-off-by: Octavian Purdila <octavian.purdila@intel.com>
Signed-off-by: Du Chenyang <chenyang.du@intel.com>
Singed-off-by: Shuo Gao <shuo.gao@intel.com>
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>
Signed-off-by: Jack Ren <jack.ren@intel.com>
Author-tracking-BZ:27414, OTC-Bug:ANDROID-835