Remove ubrk_setText hack.
This was to work around a bug in WebKit. Newer WebKit releases don't
have this bug (fixed by http://trac.webkit.org/changeset/109144), and
our ancient external/webkit doesn't have it either --- both calls to
ubrk_setText come after a null check.

I tested with the HTML in the bug shown in a WebView, with and without
this patch.

Bug: http://bugs.icu-project.org/trac/ticket/9115
Bug: http://code.google.com/p/chromium/issues/detail?id=95486
Change-Id: I7b8dc38f933eff124af02cb911e188aaac046554