Update patch set 1
Patch Set 1: Code-Review-1

Craig is waiting this changes so he can run through the ICU AlphabeticIndex tests.  We decided to make the change in one CL. His patch for the enhancement for AlphabeticIndex will include the updated ko.txt and icu*.dat.
I will abandon this CL.

Thanks for the instructions!

Patch-set: 1