Update patch set 1
Patch Set 1:

Yes, this is ready for submit ASFAIK.  This includes the change that Claire was making in ko.txt.

There are some data changes for reordering scripts that are in list that I'll be uploading, I hope, later this afternoon.

Patch-set: 1
Reviewer: Craig Cornelius <1015078@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0