Update patch set 1
Patch Set 1:

From my understanding (see [1]) the files in question are not shipped by upstream icu4c, so the Android repo seems like the right place to fix this.

(On an unrelated note: Even if it was part of upstream, wouldn't it be better if you - having worked with them before - would submit it upstream compared to me who has no idea of the icu4c code submission practices? This is at least how any Linux distro works.)

[1] http://source.icu-project.org/repos/icu/icu/trunk/source/stubdata/

Patch-set: 1
Label: Verified=0