Update patch set 1
Patch Set 1:

Markus,

thanks for the pointers. The base UCM file for my .cnv file is linked in the commit message (I didn't find the ICU charset repo and thus Googled around until I found the one on the Apple page - I'll update the commit message). The Apple one is identical to the current upstream one, so we are in agreement about the actual change :)

The problem is that the .cnv file (which isn't managed upstream, right?) seems outdated as it still references the lower-case cedilla (0x00e7) instead of the upper-case one (0x00c7). You can check the 200th byte in the .cnv file to verify that. My patch updates the .cnv and .dat files to match the latest UCM.

The actual high level problem is that entering a lower case cedilla in an SMS results in the receiver (if non-Android) seeing an upper case cedilla instead. This (wrong) behaviour is what initially triggered this patch. See [1] for the background.

[1] http://review.cyanogenmod.org/#/c/17583/

Patch-set: 1