Update patch set 4
Patch Set 4:

if you look in the script, we currently manually add gsm-03.38-2000 and a handful of other converters to our .dat file. if we're not using them, we should remove them and remove the support from the script. the closer we are to upstream icu4c the better.

Patch-set: 4