Update patch set 4
Patch Set 4:

Elliott,
I understand that this file is not official part of the SDK. However, I see apps using it an even more valid use of the SDK than e.g. the problem in [1]. When removing it, apps that worked before will now crash, which is basically the same thing as happened with the Facebook app in [1]. Please tell me if I'm missing a difference between [1] and this discussion here, but I'm not sure checking whether this is used inside the AOSP code is sufficient to tell whether it can be removed.
Having said that, I'll migrate our code to use the GsmAlphabet, which was the java code you were refering to above, right?

Craig,
this change does not add the gsm-03.38-2000.cnv file. It just fixes a bug in it, where the cedilla wasn't encoded properly (as described above).

[1] https://android.googlesource.com/platform/libcore/+/81abb6fb7332dfe62ff596ffb250d8aec61895df

Patch-set: 4