Update patch set 6
Patch Set 6:

> I'm happy to replace the "RS." with the rupee symbol since we have it in the Android font (how to verify?)

Dalvik Explorer?

> Should I revert to to "₹"?

i'm conflicted. on the one hand, it's what we're already shipping with Android so this might seem like a regression. on the other hand, i did kind of like the idea of being exactly the same as Google's web properties. can't win!

seems like this is something we can fix later though. i'd still like to get everything committed tomorrow if we can. looks like the native tests are all passing and it's just the libcore tests left? how are they looking?

Patch-set: 6
Label: Verified=0