Update patch set 1
Patch Set 1:

no, i was just telling him that i didn't accidentally miss this. now we're on the google3 patches to the code, we can presumably take the google3 patches to the tests instead of rolling our own. but for now i just wanted to make sure we're passing the relevant icu4c euc-jp tests.

Patch-set: 1
Label: SUBM=+1