Update patch set 1
Patch Set 1: Code-Review+1

It was too late to get this into ICU 52. It is also not in the trunk because I am hoping to replace the hardcoded list of first-character-in-each-script in AlphabeticIndex with a better solution based on the new collation code and the new data. If in January that looks doubtful, I should then put this patch into ICU 53 and port it to Java.

Patch-set: 1
Reviewer: Markus Scherer <1022037@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1