[Opinion needed] Fix minusSign for arabic locales.
The minusSign constants for arabic (and some other locales) have
RtL and LtR marks \u200f and \u200e). Is this correct / desirable ?

Seems very strange to include it. This causes several bugs on
android because we assume minusSign is a single character. I could
just as easily remove that assumption (though it's complicated by
the fact that DecimalFormatSymbols is a public API).

This seems to have been a deliberate change, given that there's a
unit test that tests it.

Change-Id: Ibfd5762531bd9f972b8fda0d7dda2dab8d93c05b