Update patch set 1
Patch Set 1:

(2 comments)

Ah, the assumption of a single code point explains it. Thanks for investigating.  We should eliminate that assumption if possible.

The RTL and LTR markers were added so that text around a formatted integers works correctly. It was intentional. Changing the output of DecimalFormatSymbols isn't ideal, but it should make life slightly better for Arabic users if we keep it.

Patch-set: 1
Reviewer: Craig Cornelius <1015078@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0