Modify date and time format according to local conventions
These modifications regulate the date and time format in de.txt,
es.txt, es_US.txt and th.txt.

The four files correspond to 4 locales: German (Germany),
Spanish (Spain), Spanish (United States) and Thai (Thailand).

For German (Germany):

We keep the digit date format to show the same performance

For Spanish (Spain):

We want to make it more suitable for local culture (for Spanish
(Spain) "de" is written between day and month or between month and
year), and have the same performance

For Spanish (United States):

We keep the digit date format to show the same performance

As Spanish (Spain), we want to make it more suitable for local
culture (for Spanish (United States) "de" is written between day
and month or between month and year), and have the same performance

For Thai (Thailand):

We regulate the wrong date format, when you use a date format with
weekday (like the date format have year, month, day and weekday not
just weekday), you should be careful. You should write an extra char
after weekday name

And we make it more suitable for local culture, in Thai (Thailand),
Buddhist is more common used. So when you use the Gregorian, you'd
better add the era symbol

Change-Id: Iee3954ee0dde0698820c5689fd399424ccf1bc27